Increase the bank size to 18 bits from 8 since we are now using the io expander!

+++ Chipset/src/main.cc
@@ -110,7 +110,7 @@ setupDevices() noexcept {
 [[gnu::always_inline]]
 inline
 void 
-setBankIndex(uint8_t value) {
+setBankIndex(uint32_t value) {
     AddressLinesInterface.bankSwitching.bank = value;
 }
